// packages/schemas/src/openapi/3.1/security-scheme.ts
import { intersection, literal, object, optional, string, union } from '@scalar/validation'

import { XDefaultScopes } from '@/extensions/security'

import { openApiOAuthFlowsObject } from './oauth'

const securitySchemeBase = object({
  description: optional(
    string({
      typeComment: 'A description for security scheme. CommonMark syntax MAY be used for rich text representation.',
    }),
  ),
})

const apiKeySecurityScheme = object(
  {
    ...securitySchemeBase.properties,
    type: literal('apiKey'),
    name: string({ typeComment: 'REQUIRED. The name of the header, query or cookie parameter to be used.' }),
    in: union([literal('query'), literal('header'), literal('cookie')], {
      typeComment: 'REQUIRED. The location of the API key. Valid values are "query", "header", or "cookie".',
    }),
  },
  { typeName: 'ApiKeySecuritySchemeObject' },
)

const httpSecurityScheme = object(
  {
    ...securitySchemeBase.properties,
    type: literal('http'),
    scheme: union([literal('basic'), literal('bearer')], {
      typeName: 'HttpSecuritySchemeScheme',
      typeComment:
        'REQUIRED. The name of the HTTP Authentication scheme to be used in the Authorization header as defined in RFC7235.',
    }),
    bearerFormat: optional(
      string({
        typeComment:
          'A hint to the client to identify how the bearer token is formatted. Bearer tokens are usually generated by an authorization server, so this information is primarily for documentation purposes.',
      }),
    ),
  },
  { typeName: 'HttpSecuritySchemeObject' },
)

const oauth2SecurityScheme = intersection(
  [
    object({
      ...securitySchemeBase.properties,
      type: literal('oauth2'),
      flows: openApiOAuthFlowsObject,
    }),
    XDefaultScopes,
  ],
  { typeName: 'OAuth2SecuritySchemeObject' },
)

const openIdConnectSecurityScheme = object(
  {
    ...securitySchemeBase.properties,
    type: literal('openIdConnect'),
    openIdConnectUrl: string({
      typeComment: 'REQUIRED. Well-known URL to discover the [[OpenID-Connect-Discovery]] provider metadata.',
    }),
  },
  { typeName: 'OpenIdConnectSecuritySchemeObject' },
)

export const openApiSecuritySchemeObject = union(
  [apiKeySecurityScheme, httpSecurityScheme, oauth2SecurityScheme, openIdConnectSecurityScheme],
  { typeName: 'SecuritySchemeObject' },
)